Graphe 2eme essai, parcours pas codé

+++ Pacman/src/en/master/Timer.java
@@ -3,15 +3,15 @@ package en.master;
 public class Timer {
 
 	// 1/FPS pour savoir s'il faut faire l'update
-	public static final long FPS = 60;
+	public static final short FPS = 60;
 	// GMVPS/FPS pour savoir s'il faut faire bouger le fantome
-	public static final long GMVPS = 5;
-	public static final long VMVPS = 2;
+	public static final short GMVPS = 5;
+	public static final short VMVPS = 2;
 	// PMVPS/FPS pour savoir s'il faut faire bouger Pacman
-	public static final long PMVPS = 4;
+	public static final short PMVPS = 4;
 	// Le temps en secondes que les fantomes restent en prison au debut du jeu
-	public static final long PRISON = 8;
+	public static final short PRISON = 8;
 	// La durée d'efficacité d'une supergum
-	public static final long SUPERGUM = 7;
+	public static final short SUPERGUM = 7;
 
 }
